--- src/main/java/com/music_player_client/music_player_Rest_Client/config/RestClientConfig.java
@@ -1,14 +1,0 @@
-package com.music_player_client.music_player_Rest_Client.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.client.RestTemplate;
-
-@Configuration
-public class RestClientConfig {
-
-    @Bean
-    public RestTemplate getRestTemplate() {
-        return new RestTemplate();
-    }
-}
