--- src/main/java/com/music_player_client/music_player_Rest_Client/xexperimtntal/jsonParser/entity/Menu.java
@@ -1,17 +1,0 @@
-package com.music_player_client.music_player_Rest_Client.xexperimtntal.jsonParser.entity;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class Menu {
-
-    private String menuId;
-    private String menuName;
-    private String rootCategoryId;
-    private String createdAt;
-    private String updatedAt;
-}
