+++ src/main/java/com/music_player_client/music_player_rest_client/config/RestClientConfig.java
@@ -1,0 +1,14 @@
+package com.music_player_client.music_player_rest_client.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.client.RestTemplate;
+
+@Configuration
+public class RestClientConfig {
+
+    @Bean
+    public RestTemplate getRestTemplate() {
+        return new RestTemplate();
+    }
+}
