+++ src/main/java/com/music_player_client/music_player_rest_client/xexperimtntal/jsonParser/entity/ChildCategory.java
@@ -1,0 +1,13 @@
+package com.music_player_client.music_player_rest_client.xexperimtntal.jsonParser.entity;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class ChildCategory {
+
+    private String categoryId;
+}
