+++ src/main/java/com/music_player_client/music_player_rest_client/xexperimtntal/jsonParser/entity/ChildProduct.java
@@ -1,0 +1,13 @@
+package com.music_player_client.music_player_rest_client.xexperimtntal.jsonParser.entity;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class ChildProduct {
+
+    private String productId;
+}
