--- src/test/java/com/music_player_client/music_player_Rest_Client/EmbeddedDBConfig/EmbeddedHsqlDBConfig.java
@@ -1,32 +1,0 @@
-package com.music_player_client.music_player_Rest_Client.EmbeddedDBConfig;
-
-import org.hsqldb.util.DatabaseManagerSwing;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Profile;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.datasource.embedded.EmbeddedDatabase;
-import org.springframework.jdbc.datasource.embedded.EmbeddedDatabaseBuilder;
-import org.springframework.jdbc.datasource.embedded.EmbeddedDatabaseType;
-
-import javax.sql.DataSource;
-
-@Configuration
-@Profile({"test"})
-public class EmbeddedHsqlDBConfig {
-
-    @Bean
-    public DataSource dataSource() {
-        EmbeddedDatabaseBuilder builder = new EmbeddedDatabaseBuilder();
-        EmbeddedDatabase databaseBuilder
-                = builder
-                .setType(EmbeddedDatabaseType.HSQL)
-                .addScript("classpath:tableScript.sql")
-                .build();
-        return databaseBuilder;
-    }
-    @Bean
-    public JdbcTemplate getJdbcTemplate(){
-        return new JdbcTemplate(dataSource());
-    }
-}
